+++ packages/backend-core/src/lib/index.ts
@@ -3,6 +3,5 @@ export * from './money-price-helper';
 export * from './fingerprint-config';
 export * from './credit-init';
 export * from './stripe-config';
-export * from './auth-utils';
 export * from './upstash-config';
 export * from './upstash';